Adopt consistent naming scheme for patterns
