+++ src/main/java/edu/kit/compiler/codegen/pattern/BinaryInstructionPattern.java
@@ -18,24 +18,28 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class UnaryInstruction implements Pattern<InstructionMatch> {
+public class BinaryInstructionPattern implements Pattern<InstructionMatch> {
 
     private final ir_opcode opcode;
     private final String command;
-    private final Pattern<? extends OperandMatch<? extends Operand.Destination>> operand;
+    private final Pattern<? extends OperandMatch<? extends Operand.Destination>> left;
+    private final Pattern<OperandMatch<Operand.Register>> right;
     private final boolean overwritesRegister;
     private final boolean hasMemory;
 
     @Override
     public InstructionMatch match(Node node, MatcherState matcher) {
         if (node.getOpCode() == opcode) {
-            assert node.getPredCount() == 1 + getOffset();
-            var match = operand.match(node.getPred(getOffset()), matcher);
+            var offset = hasMemory ? 1 : 0;
 
-            if (match.matches()) {
-                var mode = match.getOperand().getMode();
+            assert node.getPredCount() == 2 + offset;
+            var lhs = left.match(node.getPred(offset), matcher);
+            var rhs = right.match(node.getPred(offset + 1), matcher);
+
+            if (lhs.matches() && rhs.matches()) {
+                var mode = getMode(node);
                 var destination = getDestination(matcher::getNewRegister);
-                return new UnaryInstructionMatch(node, match, destination, mode);
+                return new BinaryInstructionMatch(node, lhs, rhs, destination, mode);
             } else {
                 return InstructionMatch.none();
             }
@@ -44,19 +48,26 @@ public class UnaryInstruction implements Pattern<InstructionMatch> {
         }
     }
 
-    private int getOffset() {
-        return hasMemory ? 1 : 0;
-    }
-
     private Optional<Integer> getDestination(Supplier<Integer> register) {
         return overwritesRegister ? Optional.of(register.get()) : Optional.empty();
     }
 
+    private Mode getMode(Node node) {
+        return switch (node.getOpCode()) {
+            case iro_Store -> {
+                var store = (firm.nodes.Store) node;
+                yield store.getType().getMode();
+            }
+            default -> node.getMode();
+        };
+    }
+
     @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-    public final class UnaryInstructionMatch extends InstructionMatch.Basic {
+    public final class BinaryInstructionMatch extends InstructionMatch.Basic {
 
         private final Node node;
-        private final OperandMatch<? extends Operand.Destination> match;
+        private final OperandMatch<? extends Operand.Destination> left;
+        private final OperandMatch<Operand.Register> right;
         private final Optional<Integer> destination;
         private final Mode mode;
 
@@ -81,7 +92,7 @@ public class UnaryInstruction implements Pattern<InstructionMatch> {
 
         @Override
         public Stream<Node> getPredecessors() {
-            var preds = match.getPredecessors();
+            var preds = Stream.concat(left.getPredecessors(), right.getPredecessors());
             if (hasMemory) {
                 preds = Stream.concat(Stream.of(node.getPred(0)), preds);
             }
@@ -92,10 +103,10 @@ public class UnaryInstruction implements Pattern<InstructionMatch> {
         private Instruction getAsOperation() {
             assert destination.isPresent();
 
-            var target = Operand.register(match.getOperand().getMode(), destination.get());
+            var target = Operand.register(right.getOperand().getMode(), destination.get());
 
-            var inputRegisters = new ArrayList<>(match.getOperand().getSourceRegisters());
-            var overwriteRegister = match.getOperand().getDestinationRegister();
+            var inputRegisters = getInputRegisters();
+            var overwriteRegister = left.getOperand().getDestinationRegister();
 
             // make sure the overwritten register is not part of input registers
             if (overwriteRegister.isPresent()) {
@@ -103,7 +114,7 @@ public class UnaryInstruction implements Pattern<InstructionMatch> {
             }
 
             return Instruction.newOp(
-                    Util.formatCmd(command, Util.getSize(mode), target),
+                    Util.formatCmd(command, Util.getSize(mode), right.getOperand(), target),
                     inputRegisters, overwriteRegister, destination.get());
         }
 
@@ -111,8 +122,15 @@ public class UnaryInstruction implements Pattern<InstructionMatch> {
             assert !destination.isPresent();
 
             return Instruction.newInput(
-                    Util.formatCmd(command, Util.getSize(mode), match.getOperand()),
-                    match.getOperand().getSourceRegisters());
+                    Util.formatCmd(command, Util.getSize(mode), right.getOperand(), left.getOperand()),
+                    getInputRegisters());
+        }
+
+        private List<Integer> getInputRegisters() {
+            var input = new ArrayList<>(left.getOperand().getSourceRegisters());
+            input.addAll(right.getOperand().getSourceRegisters());
+
+            return input;
         }
     }
 }
