+++ src/main/java/edu/kit/compiler/codegen/pattern/UnaryInstructionPattern.java
@@ -18,28 +18,24 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class BinaryInstruction implements Pattern<InstructionMatch> {
+public class UnaryInstructionPattern implements Pattern<InstructionMatch> {
 
     private final ir_opcode opcode;
     private final String command;
-    private final Pattern<? extends OperandMatch<? extends Operand.Destination>> left;
-    private final Pattern<OperandMatch<Operand.Register>> right;
+    private final Pattern<? extends OperandMatch<? extends Operand.Destination>> operand;
     private final boolean overwritesRegister;
     private final boolean hasMemory;
 
     @Override
     public InstructionMatch match(Node node, MatcherState matcher) {
         if (node.getOpCode() == opcode) {
-            var offset = hasMemory ? 1 : 0;
+            assert node.getPredCount() == 1 + getOffset();
+            var match = operand.match(node.getPred(getOffset()), matcher);
 
-            assert node.getPredCount() == 2 + offset;
-            var lhs = left.match(node.getPred(offset), matcher);
-            var rhs = right.match(node.getPred(offset + 1), matcher);
-
-            if (lhs.matches() && rhs.matches()) {
-                var mode = getMode(node);
+            if (match.matches()) {
+                var mode = match.getOperand().getMode();
                 var destination = getDestination(matcher::getNewRegister);
-                return new BinaryInstructionMatch(node, lhs, rhs, destination, mode);
+                return new UnaryInstructionMatch(node, match, destination, mode);
             } else {
                 return InstructionMatch.none();
             }
@@ -48,26 +44,19 @@ public class BinaryInstruction implements Pattern<InstructionMatch> {
         }
     }
 
+    private int getOffset() {
+        return hasMemory ? 1 : 0;
+    }
+
     private Optional<Integer> getDestination(Supplier<Integer> register) {
         return overwritesRegister ? Optional.of(register.get()) : Optional.empty();
     }
 
-    private Mode getMode(Node node) {
-        return switch (node.getOpCode()) {
-            case iro_Store -> {
-                var store = (firm.nodes.Store) node;
-                yield store.getType().getMode();
-            }
-            default -> node.getMode();
-        };
-    }
-
     @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-    public final class BinaryInstructionMatch extends InstructionMatch.Basic {
+    public final class UnaryInstructionMatch extends InstructionMatch.Basic {
 
         private final Node node;
-        private final OperandMatch<? extends Operand.Destination> left;
-        private final OperandMatch<Operand.Register> right;
+        private final OperandMatch<? extends Operand.Destination> match;
         private final Optional<Integer> destination;
         private final Mode mode;
 
@@ -92,7 +81,7 @@ public class BinaryInstruction implements Pattern<InstructionMatch> {
 
         @Override
         public Stream<Node> getPredecessors() {
-            var preds = Stream.concat(left.getPredecessors(), right.getPredecessors());
+            var preds = match.getPredecessors();
             if (hasMemory) {
                 preds = Stream.concat(Stream.of(node.getPred(0)), preds);
             }
@@ -103,10 +92,10 @@ public class BinaryInstruction implements Pattern<InstructionMatch> {
         private Instruction getAsOperation() {
             assert destination.isPresent();
 
-            var target = Operand.register(right.getOperand().getMode(), destination.get());
+            var target = Operand.register(match.getOperand().getMode(), destination.get());
 
-            var inputRegisters = getInputRegisters();
-            var overwriteRegister = left.getOperand().getDestinationRegister();
+            var inputRegisters = new ArrayList<>(match.getOperand().getSourceRegisters());
+            var overwriteRegister = match.getOperand().getDestinationRegister();
 
             // make sure the overwritten register is not part of input registers
             if (overwriteRegister.isPresent()) {
@@ -114,7 +103,7 @@ public class BinaryInstruction implements Pattern<InstructionMatch> {
             }
 
             return Instruction.newOp(
-                    Util.formatCmd(command, Util.getSize(mode), right.getOperand(), target),
+                    Util.formatCmd(command, Util.getSize(mode), target),
                     inputRegisters, overwriteRegister, destination.get());
         }
 
@@ -122,15 +111,8 @@ public class BinaryInstruction implements Pattern<InstructionMatch> {
             assert !destination.isPresent();
 
             return Instruction.newInput(
-                    Util.formatCmd(command, Util.getSize(mode), right.getOperand(), left.getOperand()),
-                    getInputRegisters());
-        }
-
-        private List<Integer> getInputRegisters() {
-            var input = new ArrayList<>(left.getOperand().getSourceRegisters());
-            input.addAll(right.getOperand().getSourceRegisters());
-
-            return input;
+                    Util.formatCmd(command, Util.getSize(mode), match.getOperand()),
+                    match.getOperand().getSourceRegisters());
         }
     }
 }
